XhyveDarwinPkg: add Apple boot support
Apple's boot.efi checks if the ConsoleControl protocol returns
EFI_SUCCESS in both GetMode and SetMode.

Apple uses a kernel extension to make parts of the DataHub protocol
available to the xnu kernel. The xnu kernel then checks for FSB
frequency and if it's not found fails with:
"tsc_init: EFI not supported!"

Lastly, some firmware settings are added to AppleSupport and written to
nvram.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Reza Jelveh <reza.jelveh@tuhh.de>
Signed-off-by: Vladislav Yaroshchuk <yaroshchuk2000@gmail.com>